fix list item navigation
